fix extract prototypes multi

+++ Maps.py
@@ -27,11 +27,12 @@ def pix2meter(args):
     # values should be stored accordingly
     # 1. reprojected tracks: args.ReprojectedPkl -> args.ReprojectedPklMeter
     r = meter_per_pixel(args.MetaData['center'])
-    df = pd.read_pickle(args.ReprojectedPkl)
-    scale_function = lambda x: r*x
-    df['x'] = df['x'].apply(scale_function)
-    df['y'] = df['y'].apply(scale_function)
-    df.to_pickle(args.ReprojectedPklMeter)
+    if os.path.exists(args.ReprojectedPkl):
+        df = pd.read_pickle(args.ReprojectedPkl)
+        scale_function = lambda x: r*x
+        df['x'] = df['x'].apply(scale_function)
+        df['y'] = df['y'].apply(scale_function)
+        df.to_pickle(args.ReprojectedPklMeter)
 
     # 2. labeled trackes: args.TrackLabellingExportPth -> args.TrackLabellingExportPthMeter
     if os.path.exists(args.TrackLabellingExportPth):
@@ -39,6 +40,9 @@ def pix2meter(args):
         scale_function = lambda x: np.array([[r*xi[0],r*xi[1]] for xi in x])
         df['trajectory'] = df['trajectory'].apply(scale_function)
         df.to_pickle(args.TrackLabellingExportPthMeter)
+    else:
+        print(args.TrackLabellingExportPth)
+        raise "What the fuck"
 
     # # show the scale with roi
     # print(r)
